docs: update conda best practices (#363)
Now that mamba has been absorbed into mainline conda, the installation instructions can be simpler. I noticed this while trying to get students started: I don't want to have to say, "ignore all the mamba stuff," as that just makes it more complicated.

I also added a CUDA section as a drop-down. We're not even close to having simple recipes that anyone can follow to install packages that depend on CUDA, but at least we can get them started with some pointers. For instance, it's helpful to let people know that CUDA itself (at the time of this writing) cannot be installed as a conda package. (The packages that claim to do this are _not_ robust.)